API: Rename Panel to TogglePanel and move to widgets
