+++ typhon/panel.py
@@ -6,9 +6,7 @@ import logging
 ############
 # External #
 ############
-from pydm.PyQt.QtCore import pyqtSlot
 from pydm.PyQt.QtGui import QHBoxLayout, QFont, QLabel, QWidget, QGridLayout
-from pydm.PyQt.QtGui import QPushButton, QVBoxLayout
 from pydm.widgets import PyDMLineEdit
 
 #############
@@ -20,71 +18,7 @@ from .widgets import TyphonComboBox, TyphonLabel
 logger = logging.getLogger(__name__)
 
 
-class Panel(QWidget):
-    """
-    Generic Panel Widget
-
-    Displays a widget below QPushButton that hides and shows the contents. It
-    is up to subclasses to re-point the attribute :attr:`.contents` to the
-    widget whose visibility you would like to toggle.
-
-    By default, it is assumed that the Panel is initialized with the
-    :attr:`.contents` widget as visible, however the contents will be hidden
-    and the button synced to the proper position if :meth:`.show_contents` is
-    called after instance creation
-
-    Parameters
-    ----------
-    title : str
-        Title of Panel. This will be the text on the QPushButton
-
-    parent : QWidget
-
-    Attributes
-    ----------
-    contents : QWidget
-        Widget whose visibility is controlled via the QPushButton
-    """
-    def __init__(self, title, parent=None):
-        super().__init__(parent=parent)
-        # Create Widget Infrastructure
-        self.title = title
-        self.setLayout(QVBoxLayout())
-        self.layout().setContentsMargins(2, 2, 2, 2)
-        self.layout().setSpacing(5)
-        # Create button control
-        # Assuming widget is visible, set the button as checked
-        self.contents = None
-        self.hide_button = QPushButton(self.title)
-        self.hide_button.setCheckable(True)
-        self.hide_button.setChecked(True)
-        self.layout().addWidget(self.hide_button)
-        self.hide_button.clicked.connect(self.show_contents)
-
-    @pyqtSlot(bool)
-    def show_contents(self, show):
-        """
-        Show the contents of the Widget
-
-        Hides the :attr:`.contents` QWidget and sets the :attr:`.hide_button`
-        to the proper status to indicate whether the widget is hidden or not
-
-        Parameters
-        ----------
-        show : bool
-        """
-        # Configure our button in case this slot was called elsewhere
-        self.hide_button.setChecked(show)
-        # Show or hide the widget if the contents exist
-        if self.contents:
-            if show:
-                self.show()
-                self.contents.show()
-            else:
-                self.contents.hide()
-
-
-class SignalPanel(Panel):
+class SignalPanel(QWidget):
     """
     Base panel display for EPICS signals
