+++ typhon/widgets.py
@@ -8,8 +8,9 @@ import os.path
 # External #
 ############
 from pydm.PyQt import uic
-from pydm.PyQt.QtCore import QSize, Qt
-from pydm.PyQt.QtGui import QStackedWidget, QLabel, QAbstractButton
+from pydm.PyQt.QtCore import QSize, Qt, pyqtSlot
+from pydm.PyQt.QtGui import QAbstractButton, QStackedWidget, QLabel
+from pydm.PyQt.QtGui import QVBoxLayout, QPushButton, QWidget
 from pydm.widgets import PyDMDrawingImage, PyDMLabel, PyDMEnumComboBox
 
 ###########
@@ -18,6 +19,70 @@ from pydm.widgets import PyDMDrawingImage, PyDMLabel, PyDMEnumComboBox
 from .utils import ui_dir, channel_name
 
 logger = logging.getLogger(__name__)
+
+
+class TogglePanel(QWidget):
+    """
+    Generic Panel Widget
+
+    Displays a widget below QPushButton that hides and shows the contents. It
+    is up to subclasses to re-point the attribute :attr:`.contents` to the
+    widget whose visibility you would like to toggle.
+
+    By default, it is assumed that the Panel is initialized with the
+    :attr:`.contents` widget as visible, however the contents will be hidden
+    and the button synced to the proper position if :meth:`.show_contents` is
+    called after instance creation
+
+    Parameters
+    ----------
+    title : str
+        Title of Panel. This will be the text on the QPushButton
+
+    parent : QWidget
+
+    Attributes
+    ----------
+    contents : QWidget
+        Widget whose visibility is controlled via the QPushButton
+    """
+    def __init__(self, title, parent=None):
+        super().__init__(parent=parent)
+        # Create Widget Infrastructure
+        self.title = title
+        self.setLayout(QVBoxLayout())
+        self.layout().setContentsMargins(2, 2, 2, 2)
+        self.layout().setSpacing(5)
+        # Create button control
+        # Assuming widget is visible, set the button as checked
+        self.contents = None
+        self.hide_button = QPushButton(self.title)
+        self.hide_button.setCheckable(True)
+        self.hide_button.setChecked(True)
+        self.layout().addWidget(self.hide_button)
+        self.hide_button.clicked.connect(self.show_contents)
+
+    @pyqtSlot(bool)
+    def show_contents(self, show):
+        """
+        Show the contents of the Widget
+
+        Hides the :attr:`.contents` QWidget and sets the :attr:`.hide_button`
+        to the proper status to indicate whether the widget is hidden or not
+
+        Parameters
+        ----------
+        show : bool
+        """
+        # Configure our button in case this slot was called elsewhere
+        self.hide_button.setChecked(show)
+        # Show or hide the widget if the contents exist
+        if self.contents:
+            if show:
+                self.show()
+                self.contents.show()
+            else:
+                self.contents.hide()
 
 
 class TyphonComboBox(PyDMEnumComboBox):
